Refactor Edit component code-behind
